+++ Source/Function Root Finder Algorithms/secant.c
@@ -90,7 +90,7 @@ int main() {
     } // end of if mode
 
     // calculation
-    double x = secant(expression, a0, b0, ete, ere, tol, maxiter, mode, &flag);
+    double x = secant(expression, a0, b0, ete, ere, tol, (unsigned int) maxiter, mode, &flag);
 
     // if there was an answer
     if (flag) {
